bff/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
